chore(deps): migrate storybook configuration to ESM and update dependencies
