--- apps/public/.storybook/main.js
@@ -1,27 +1,0 @@
-import { dirname, join } from "path";
-module.exports = {
-  stories: ["../src/**/*.stories.@(js|jsx|ts|tsx|mdx)"],
-
-  "addons": [
-    getAbsolutePath("@storybook/addon-links"),
-    getAbsolutePath("@storybook/preset-create-react-app"),
-    "@storybook/addon-essentials",
-    "@storybook/addon-interactions",
-    "@storybook/addon-a11y",
-    "@chromatic-com/storybook"
-  ],
-
-  "framework": {
-    name: "@storybook/react-webpack5",
-    options: {}
-  },
-  staticDirs: ["../public"],
-
-  docs: {
-    autodocs: true
-  }
-}
-
-function getAbsolutePath(value) {
-  return dirname(require.resolve(join(value, "package.json")));
-}
