+++ apps/public/package.json
@@ -82,7 +82,6 @@
     "util": "^0.12.5",
     "vm-browserify": "^1.1.2",
     "web-vitals": "^2.1.4",
-    "webpack-dev-server": "^5.2.5",
     "webpack-manifest-plugin": "^4.0.2",
     "workbox-webpack-plugin": "^6.4.1",
     "zustand": "^4.1.1"
@@ -126,18 +125,13 @@
     ]
   },
   "devDependencies": {
-    "@chromatic-com/storybook": "^3.2.7",
+    "@chromatic-com/storybook": "^5.2.1",
     "@rushstack/eslint-patch": "^1.15.0",
-    "@storybook/addon-a11y": "8.6.15",
-    "@storybook/addon-actions": "8.6.15",
-    "@storybook/addon-essentials": "8.6.15",
-    "@storybook/addon-interactions": "8.6.15",
-    "@storybook/addon-links": "8.6.15",
-    "@storybook/node-logger": "8.6.15",
-    "@storybook/preset-create-react-app": "8.6.15",
-    "@storybook/react": "8.6.15",
-    "@storybook/react-webpack5": "^8.6.15",
-    "@storybook/test": "^8.6.15",
+    "@storybook/addon-a11y": "10.5.3",
+    "@storybook/addon-links": "10.5.3",
+    "@storybook/preset-create-react-app": "10.5.3",
+    "@storybook/react": "10.5.3",
+    "@storybook/react-webpack5": "10.5.3",
     "chromatic": "^6.5.4",
     "confusing-browser-globals": "^1.0.11",
     "eslint-plugin-import": "^2.32.0",
@@ -145,6 +139,7 @@
     "eslint-plugin-react": "^7.37.5",
     "eslint-plugin-react-hooks": "^4.6.2",
     "natural-compare-lite": "^1.4.0",
+    "storybook": "10.5.3",
     "string-natural-compare": "^3.0.1",
     "tsutils": "^3.21.0",
     "typescript": "^5.1.6",
